Move capitalize, isCapitalized to a new module
